% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDIA_functions_for_win_and_mac.R
\name{call_mapDIA_windows}
\alias{call_mapDIA_windows}
\title{this function actually calls mapDIA and takes care of windows pathnames}
\usage{
call_mapDIA_windows()
}
\description{
this function actually calls mapDIA and takes care of windows pathnames
}
