% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_Frag_functions_for_win_and_mac.R
\name{call_saintQf_windows}
\alias{call_saintQf_windows}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine}
\usage{
call_saintQf_windows()
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine
}
