% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_Prot_functions_for_win_and_mac.R
\name{call_saintQp_mac}
\alias{call_saintQp_mac}
\title{this function actually calls saintQprotein and takes care of mac pathnames}
\usage{
call_saintQp_mac()
}
\description{
this function actually calls saintQprotein and takes care of mac pathnames
}
