% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_PROT_SC_prepData_callFunction.R
\name{call_saintQP_SC_windows}
\alias{call_saintQP_SC_windows}
\title{runs SaintQ on spectral counts for windows}
\usage{
call_saintQP_SC_windows()
}
\description{
runs SaintQ on spectral counts for windows
}
