% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ms_identification_summary.R
\name{do_ms_identification_summary}
\alias{do_ms_identification_summary}
\title{this function collects summary information}
\usage{
do_ms_identification_summary(result)
}
\description{
this function collects summary information
}
