% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_raw.R
\name{ibq_clean_data}
\alias{ibq_clean_data}
\title{this function keeps the tibble structure and prepares the data for following steps
NA values are replaced by 0
it writes a file with all gene, protein, description information into the working directory
min_number_of_quantified_peps lets you choose how many peptides need to be quantified per run
to accept a protein as identidied and quantified}
\usage{
ibq_clean_data(results, min_number_of_quantified_peps, ...)
}
\description{
this function keeps the tibble structure and prepares the data for following steps
NA values are replaced by 0
it writes a file with all gene, protein, description information into the working directory
min_number_of_quantified_peps lets you choose how many peptides need to be quantified per run
to accept a protein as identidied and quantified
}
