% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_from_raw.R
\name{ibq_clean_data_keep_mods}
\alias{ibq_clean_data_keep_mods}
\title{same as the other but attaches modification information}
\usage{
ibq_clean_data_keep_mods(results, min_number_of_quantified_peps, ...)
}
\description{
same as the other but attaches modification information
}
