% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ms_identification_summary.R
\name{ibq_do_ms_identification_summary}
\alias{ibq_do_ms_identification_summary}
\title{this function collects summary information}
\usage{
ibq_do_ms_identification_summary(result)
}
\description{
this function collects summary information
}
