% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibq_deal_with_contaminant_list.R
\name{ibq_filter_contam}
\alias{ibq_filter_contam}
\title{to filter out common contaminants via gene_name or protein id
a short list is delivered with the package and generated in the working folder
it will not be generated if a list already exists
so you can use your own}
\usage{
ibq_filter_contam()
}
\description{
to filter out common contaminants via gene_name or protein id
a short list is delivered with the package and generated in the working folder
it will not be generated if a list already exists
so you can use your own
}
