% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDIA_diagonal_and_scatter_plot_ly.R
\name{ibq_format_mapDIA_for_drawing}
\alias{ibq_format_mapDIA_for_drawing}
\title{after mapDIA has been ran we can look at the results of the single pairwise
comparisons. so we need to do some data wrangling to generate the
dataframes that go into the draawings. also we set the FDR cut off.
we start form the md tibble}
\usage{
ibq_format_mapDIA_for_drawing(FDRcutoff = 0.01)
}
\description{
after mapDIA has been ran we can look at the results of the single pairwise
comparisons. so we need to do some data wrangling to generate the
dataframes that go into the draawings. also we set the FDR cut off.
we start form the md tibble
}
