% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibq_group_by_folder.R
\name{ibq_group_by_folder}
\alias{ibq_group_by_folder}
\title{this little charm of a function will help us to group stuff bein in the same folder aka in the same analysis}
\usage{
ibq_group_by_folder()
}
\description{
this little charm of a function will help us to group stuff bein in the same folder aka in the same analysis
}
