% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mapDIA_winmac.R
\name{ibq_mapDIA}
\alias{ibq_mapDIA}
\title{system.file("extdata/win", "mapDIA.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis
it should recognize the operating system automatically
combine gene_sequence option to true will treat peptides as proteins
per definition the binary comparisons are made with the alphabetically last
sample name so you can influence it via the name.
system.file("extdata/win", "mapDIA.exe", package = "ibqreader")}
\usage{
ibq_mapDIA(combine_gene_sequence)
}
\description{
system.file("extdata/win", "mapDIA.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis
it should recognize the operating system automatically
combine gene_sequence option to true will treat peptides as proteins
per definition the binary comparisons are made with the alphabetically last
sample name so you can influence it via the name.
system.file("extdata/win", "mapDIA.exe", package = "ibqreader")
}
