% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDIA_diagonal_and_scatter_plot_ly.R
\name{ibq_mapDIA_draw_diagonal}
\alias{ibq_mapDIA_draw_diagonal}
\title{mapDIA results that have been formatted with the function
ibq_format_mapDIA_for_drawing can be visualized as an interactiv plot_ly
graph with this function}
\usage{
ibq_mapDIA_draw_diagonal(bait, base, ddata, DEFAULT_COLOR = "#1976d2",
  SIG_COLOR = "#ffeb3b")
}
\description{
mapDIA results that have been formatted with the function
ibq_format_mapDIA_for_drawing can be visualized as an interactiv plot_ly
graph with this function
}
