% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDIA_diagonal_and_scatter_plot_ly.R
\name{ibq_mapDIA_draw_volcano}
\alias{ibq_mapDIA_draw_volcano}
\title{mapDIA results that have been formatted with the function
ibq_format_mapDIA_for_drawing can be visualized as an interactiv plot_ly
graph with this function from https://www.biostars.org/p/214100/}
\usage{
ibq_mapDIA_draw_volcano(bait, base, ddata)
}
\description{
mapDIA results that have been formatted with the function
ibq_format_mapDIA_for_drawing can be visualized as an interactiv plot_ly
graph with this function from https://www.biostars.org/p/214100/
}
