% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mapDIA.R
\name{ibq_mapDIA_WIN}
\alias{ibq_mapDIA_WIN}
\title{system.file("extdata/win", "mapDIA.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis}
\usage{
ibq_mapDIA_WIN(combine_gene_sequence)
}
\description{
system.file("extdata/win", "mapDIA.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis
}
