% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{ibq_name_things}
\alias{ibq_name_things}
\title{this function names and selects TMT channels that are  specified in the pooling.csv}
\usage{
ibq_name_things(bait_channel, data)
}
\description{
this function names and selects TMT channels that are  specified in the pooling.csv
}
