% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibq_read_files.R
\name{ibq_read_results}
\alias{ibq_read_results}
\title{read files form isobarquant workflow}
\usage{
ibq_read_results(thefolder)
}
\arguments{
\item{x}{the parent folder in which the the three result files are located}
}
\value{
loads data into the global enironment of the R session alos prints the pooling files into the folders
}
\description{
read files form isobarquant workflow
}
\examples{
none yet
}
