% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{ibq_relative_intensities}
\alias{ibq_relative_intensities}
\title{this function calculates the relative TMT intensities row so that the sum is one.}
\usage{
ibq_relative_intensities(bare_peptides_or_protein_data)
}
\description{
this function calculates the relative TMT intensities row so that the sum is one.
}
