% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_PROT_SC_prepData_callFunction.R
\name{ibq_saintQ_SC}
\alias{ibq_saintQ_SC}
\title{prepare data for saintQ SC}
\usage{
ibq_saintQ_SC()
}
\arguments{
\item{saint_tib_SC}{}
}
\description{
prepare data for saintQ SC
}
