% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_Frag_prepData_callFunctions.R
\name{ibq_saintQfrag}
\alias{ibq_saintQfrag}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine}
\usage{
ibq_saintQfrag(combine_gene_sequence)
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine
}
