% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_saintQf_MAC.R
\name{ibq_saintQfrag_MAC}
\alias{ibq_saintQfrag_MAC}
\title{system.file("extdata/mac", "saintq.exe", package = "ibqreader")
this does everything to do a saintQ fragment based analysis with a mac}
\usage{
ibq_saintQfrag_MAC(combine_gene_sequence)
}
\description{
system.file("extdata/mac", "saintq.exe", package = "ibqreader")
this does everything to do a saintQ fragment based analysis with a mac
}
