% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_saintQf_WIN.R
\name{ibq_saintQfrag_WIN}
\alias{ibq_saintQfrag_WIN}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine}
\usage{
ibq_saintQfrag_WIN(combine_gene_sequence)
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a sanitQ analyis on a windows machine
}
