% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_saintQ_MAC.R
\name{ibq_saintQprot_MAC}
\alias{ibq_saintQprot_MAC}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a saintq analysis on protein level}
\usage{
ibq_saintQprot_MAC(combine_gene_sequence)
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a saintq analysis on protein level
}
