% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saintQ_Prot_prepData_callFunctions.R
\name{ibq_saintQprot_new}
\alias{ibq_saintQprot_new}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis}
\usage{
ibq_saintQprot_new(combine_gene_sequence = FALSE)
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis
}
