% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_saintQ_WIN.R
\name{ibq_saintQprot_WIN}
\alias{ibq_saintQprot_WIN}
\title{system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis}
\usage{
ibq_saintQprot_WIN(combine_gene_sequence)
}
\description{
system.file("extdata/win", "saintq.exe", package = "ibqreader")
this does everything to do a mapDIA_analysis
}
