% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{ibq_split_pepident}
\alias{ibq_split_pepident}
\title{get form fake combined unique peptide column to protein|sequence|msms_if}
\usage{
ibq_split_pepident(peptide_tibble)
}
\description{
get form fake combined unique peptide column to protein|sequence|msms_if
}
