% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gene_names.R
\name{ibq_use_gene_names}
\alias{ibq_use_gene_names}
\title{give the gene names to your tibbles
it also deals with duplicated stuff ... it will aggregate (sum for now) the single duplictes}
\usage{
ibq_use_gene_names()
}
\description{
give the gene names to your tibbles
it also deals with duplicated stuff ... it will aggregate (sum for now) the single duplictes
}
