% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_correction_factors.R
\name{ibq_apply_coorection_factors}
\alias{ibq_apply_coorection_factors}
\title{if we want to apply general correction factors, specified int the pooling.csv here we are}
\usage{
ibq_apply_coorection_factors()
}
\description{
if we want to apply general correction factors, specified int the pooling.csv here we are
}
